+++ CHANGELOG.md
@@ -50,6 +50,7 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 - improve wallpaper selection with new modal to upload a file
 - refactor HBv2 and UIv5 readiness modals
 - move backup settings from ui advanced settings to backup/restore modal
+- show lock management service info on lock mechanism info modal
 
 ### Other Changes